refactor: move the data part to Discussion.Core project
