+++ dotnetclub.sln
@@ -3,27 +3,29 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.26114.2
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Discussion.Web", "src\Discussion.Web\Discussion.Web.csproj", "{2C14F77A-A763-49AC-9751-07C10AD5E201}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Discussion.Web", "src\Discussion.Web\Discussion.Web.csproj", "{2C14F77A-A763-49AC-9751-07C10AD5E201}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{9958F5A5-400C-428F-8DC8-B0EB9DE793FD}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Discussion.Web.Tests", "test\Discussion.Web.Tests\Discussion.Web.Tests.csproj", "{6B1DCA02-EF62-42CA-8865-DA66600D84C1}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Discussion.Web.Tests", "test\Discussion.Web.Tests\Discussion.Web.Tests.csproj", "{6B1DCA02-EF62-42CA-8865-DA66600D84C1}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Discussion.Migrations", "src\Discussion.Migrations\Discussion.Migrations.csproj", "{42458253-F6AE-4176-B617-CEB185B640C6}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Discussion.Migrations", "src\Discussion.Migrations\Discussion.Migrations.csproj", "{42458253-F6AE-4176-B617-CEB185B640C6}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "_Scratches", "_Scratches", "{B724901A-4E11-4815-BA81-19D2A356E130}"
-ProjectSection(SolutionItems) = preProject
-	build-linux.sh = build-linux.sh
-	build.cake = build.cake
-	build.ps1 = build.ps1
-	build.sh = build.sh
-	.gitignore = .gitignore
-	DockerFile = DockerFile
-	deploy-on-hyper.sh = deploy-on-hyper.sh
-	upgrade-from-existing.sh = upgrade-from-existing.sh
-	readme.md = readme.md
-	readme-en.md = readme-en.md
-EndProjectSection
+	ProjectSection(SolutionItems) = preProject
+		.gitignore = .gitignore
+		build-linux.sh = build-linux.sh
+		build.cake = build.cake
+		build.ps1 = build.ps1
+		build.sh = build.sh
+		deploy-on-hyper.sh = deploy-on-hyper.sh
+		DockerFile = DockerFile
+		readme-en.md = readme-en.md
+		readme.md = readme.md
+		upgrade-from-existing.sh = upgrade-from-existing.sh
+	EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Discussion.Core", "src\Discussion.Core\Discussion.Core.csproj", "{D9E60D32-042F-4916-B006-F815E6323D12}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -43,11 +45,18 @@ Global
 		{42458253-F6AE-4176-B617-CEB185B640C6}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{42458253-F6AE-4176-B617-CEB185B640C6}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{42458253-F6AE-4176-B617-CEB185B640C6}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D9E60D32-042F-4916-B006-F815E6323D12}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D9E60D32-042F-4916-B006-F815E6323D12}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D9E60D32-042F-4916-B006-F815E6323D12}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D9E60D32-042F-4916-B006-F815E6323D12}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{6B1DCA02-EF62-42CA-8865-DA66600D84C1} = {9958F5A5-400C-428F-8DC8-B0EB9DE793FD}
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {636BCE3B-E83E-4CCF-9BA2-0386E439D9B8}
 	EndGlobalSection
 EndGlobal
