--- src/Discussion.Web/Models/Article.cs
@@ -1,13 +1,0 @@
-﻿
-using Discussion.Web.Data;
-
-namespace Discussion.Web.Models
-{
-    /// <summary>
-    /// Article 在系统中是一个基本数据类型，暂时仅用于演示和测试
-    /// </summary>
-    public class Article : Entity
-    {
-        public string Title{ get; set; }
-    }
-}
