+++ src/Discussion.Web/ViewModels/TopicListViewModel.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using Discussion.Core.Models;
 using Discussion.Web.Models;
 
 namespace Discussion.Web.ViewModels
@@ -6,7 +7,7 @@ namespace Discussion.Web.ViewModels
     public class TopicListViewModel
     {
         public List<Topic> Topics { get; set; }
-        
+
         public int CurrentPage { get; set; }
         public bool HasNextPage { get; set; }
         public bool HasPreviousPage { get; set; }
