+++ src/Discussion.Web/Views/Topic/Create.cshtml
@@ -1,36 +1,34 @@
-﻿@{
+﻿@using Discussion.Core.Models
+@{
     ViewData["Title"] = "创建新主题";
 }
 
-
-
-    <div class="row clearfix">
-        <h3>创建新话题</h3>
-        <div class="topic-title-container">
-            <input id="new-topic-title" type="text" placeholder="标题" autofocus maxlength="120" />
-        </div>
-        <div class="topic-content-container">
-            <div id="content-editor"></div>
-        </div>
+<div class="row clearfix">
+    <h3>创建新话题</h3>
+    <div class="topic-title-container">
+        <input id="new-topic-title" type="text" placeholder="标题" autofocus maxlength="120" />
     </div>
-    <div class="row topic-submit-section clearfix">
-        <button id="submit-create" class="btn btn-default submit-create-btn">创建</button>
-        <div id="topic-type-dropdown" class="dropdown topic-type-dropdown">
-            <div class="dropdown-toggle topic-type-btn" data-toggle="dropdown" aria-haspopup="true">
-                <span class="selected-type">话题类别</span>
-                <span class="caret"></span>
-            </div>
-            <ul class="dropdown-menu">
-                @foreach (var item in Html.GetEnumSelectList<TopicType>())
-                {
-                    <li class="topic-type-item"><a href="javascript:;" attr-value="@item.Value">@item.Text</a></li>
-                }
-            </ul>
-        </div>
+    <div class="topic-content-container">
+        <div id="content-editor"></div>
     </div>
+</div>
+<div class="row topic-submit-section clearfix">
+    <button id="submit-create" class="btn btn-default submit-create-btn">创建</button>
+    <div id="topic-type-dropdown" class="dropdown topic-type-dropdown">
+        <div class="dropdown-toggle topic-type-btn" data-toggle="dropdown" aria-haspopup="true">
+            <span class="selected-type">话题类别</span>
+            <span class="caret"></span>
+        </div>
+        <ul class="dropdown-menu">
+            @foreach (var item in Html.GetEnumSelectList<TopicType>())
+            {
+                <li class="topic-type-item"><a href="javascript:;" attr-value="@item.Value">@item.Text</a></li>
+            }
+        </ul>
+    </div>
+</div>
 
-
-@section Tail{    
+@section Tail{
     <script>
         window.createTopicUrl = "@Url.Action("CreateTopic")";
         window.DiscussionApp.mdeditor.setupEditor();
