+++ test/Discussion.Web.Tests/Discussion.Web.Tests.csproj
@@ -13,6 +13,7 @@
   </PropertyGroup>
   <ItemGroup>
     <ProjectReference Include="../../src/Discussion.Web/Discussion.Web.csproj" />
+    <ProjectReference Include="..\..\src\Discussion.Core\Discussion.Core.csproj" />
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="coverlet.msbuild" Version="2.3.0" />
